python_backend/requirements.txt
kafka-python>=2.0.2
pyle38


